app/pyproject.toml
[tool.poetry]
name = "oz-01-collabo-004"
version = "0.1.0"
description = ""
authors = ["hyeongbin0516 <gudqls0516@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=5.0.1,<6.0.0"
djangorestframework = ">=3.14.0,<4.0.0"
djangorestframework-simplejwt = ">=5.3.1,<6.0.0"
mysqlclient = ">=2.2.4,<3.0.0"
drf-spectacular = ">=0.27.1,<0.28.0"
uwsgi = ">=2.0.24,<2.1.0"
django-crontab = ">=0.7.1,<0.8.0"
python-dotenv = ">=1.0.1,<2.0.0"
django-cors-headers = ">=4.3.1,<5.0.0"
django-allauth = ">=0.61.1,<1.0.0"
black = ">=24.3.0,<25.0.0"
isort = "^5.13.2"

[tool.poetry.group.dev.dependencies]
mypy = "1.7.1"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
types-mysqlclient = "^2.2.0.20240311"
types-requests = "^2.31.0.20240406"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.5"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"